+++ packages/pds/src/api/app/bsky/feed/getFeed.ts
@@ -2,6 +2,8 @@ import { Server } from '../../../../lexicon'
 import AppContext from '../../../../context'
 import { pipethrough } from '../../../../pipethrough'
 import { ids } from '../../../../lexicon/lexicons'
+import { AtUri } from '@atproto/syntax'
+import { InvalidRequestError } from '@atproto/oauth-provider'
 
 export default function (server: Server, ctx: AppContext) {
   const { appViewAgent } = ctx
@@ -12,18 +14,23 @@ export default function (server: Server, ctx: AppContext) {
     handler: async ({ params, auth, req }) => {
       const requester = auth.credentials.did
 
-      const { data: feed } =
-        await appViewAgent.api.app.bsky.feed.getFeedGenerator(
-          { feed: params.feed },
-          await ctx.appviewAuthHeaders(
-            requester,
-            ids.AppBskyFeedGetFeedGenerator,
-          ),
+      const feedUrl = new AtUri(params.feed)
+      const { data } = await appViewAgent.com.atproto.repo.getRecord({
+        repo: feedUrl.hostname,
+        collection: feedUrl.collection,
+        rkey: feedUrl.rkey,
+      })
+      const feedDid = data.value['did']
+      if (typeof feedDid !== 'string') {
+        throw new InvalidRequestError(
+          'could not resolve feed did',
+          'UnknownFeed',
         )
+      }
 
       return pipethrough(ctx, req, {
         iss: requester,
-        aud: feed.view.did,
+        aud: feedDid,
         lxm: ids.AppBskyFeedGetFeedSkeleton,
       })
     },