+++ packages/oauth/oauth-client-node/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @atproto/oauth-client-node
 
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @atproto/oauth-client@0.3.1
+
 ## 0.2.0
 
 ### Minor Changes